SportsApp/requirements.txt
aioredis==2.0.1
aiosmtpd==1.4.2
aniso8601==9.0.1
asgiref==3.5.2
async-timeout==4.0.2
atpublic==3.0.1
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
defusedxml==0.7.1
Django==4.0.6
django-allauth==0.51.0
django-channels==0.7.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
graphene==3.1
graphene-django==3.0.0b7
graphql-core==3.2.1
graphql-relay==3.2.0
h2==4.1.0
hiredis==2.0.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
mysql==0.0.3
mysqlclient==2.1.1
oauthlib==3.2.0
Pillow==9.2.0
priority==2.0.0
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyngrok==5.1.0
pyOpenSSL==22.0.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
urllib3==1.26.10
websocket-client==1.3.3
zope.interface==5.4.0
